Refactor heatmap components and improve label rendering
- Extract HourlyChart into its own component file for better separation
  of concerns
- Add colLabelWidth prop to Heatmap to support labels wider than cells
- Refactor column label rendering into buildHeaderRow function using a
  cleaner while-loop approach with helper functions
- Fix yearly heatmap month labels to be evenly spaced (every ~4 weeks)
  instead of following actual calendar boundaries
