+++ src/components/HourlyChart.tsx
@@ -1,0 +1,82 @@
+import React from "react";
+import { Box, Text } from "ink";
+import type { HourlyChartProps } from "../types.js";
+import ErrorMessage from "./ErrorMessage.js";
+
+const BAR_CHARS = [" ", "▁", "▂", "▃", "▄", "▅", "▆", "▇", "█"];
+const CHART_WIDTH = 2;
+const CHART_HEIGHT = 4;
+
+function formatHour(hour: number): string {
+  const date = new Date();
+  date.setHours(hour, 0, 0, 0);
+  return date.toLocaleTimeString(undefined, {
+    hour: "numeric",
+    minute: "2-digit",
+  });
+}
+
+export default function HourlyChart(props: HourlyChartProps) {
+  const { hourlyBreakdown, peakHour, peakHourCount } = props;
+
+  const max = Math.max(...hourlyBreakdown);
+  if (max === 0) {
+    return <ErrorMessage message="No hourly data available to display." />;
+  }
+
+  // Normalize values to total height (CHART_HEIGHT rows x 8 levels per row)
+  const totalLevels = CHART_HEIGHT * 8;
+  const normalized = hourlyBreakdown.map((count) =>
+    Math.round((count / max) * totalLevels)
+  );
+
+  // Build rows from top to bottom
+  const rows: string[] = [];
+  for (let row = CHART_HEIGHT - 1; row >= 0; row--) {
+    const rowChars = normalized.map((level) => {
+      const rowBase = row * 8;
+      const inThisRow = Math.max(0, Math.min(8, level - rowBase));
+      return (BAR_CHARS[inThisRow] ?? " ").repeat(CHART_WIDTH);
+    });
+    rows.push(rowChars.join(""));
+  }
+
+  // Build peak label row
+  const peakLabel = peakHourCount.toLocaleString();
+  let labelRow = "";
+  if (peakHour !== null) {
+    const peakPos = peakHour * CHART_WIDTH;
+    const labelStart = Math.max(0, peakPos - Math.floor(peakLabel.length / 2));
+    labelRow = " ".repeat(labelStart) + peakLabel;
+  }
+
+  return (
+    <Box flexDirection="column">
+      <Text bold color="cyan">
+        Hourly Activity
+      </Text>
+      <Box flexDirection="column">
+        {peakHour !== null && (
+          <Box gap={1}>
+            <Text dimColor>Your peak hour: </Text>
+            <Text color="green">{formatHour(peakHour)}</Text>
+          </Box>
+        )}
+      </Box>
+      <Box flexDirection="column">
+        {peakHour !== null && <Text color="green">{labelRow}</Text>}
+        <Box flexDirection="column">
+          {rows.map((row, i) => (
+            <Text key={i} color="cyan">
+              {row}
+            </Text>
+          ))}
+        </Box>
+        <Box justifyContent="space-between" width={24 * CHART_WIDTH}>
+          <Text dimColor>{formatHour(0)}</Text>
+          <Text dimColor>{formatHour(24)}</Text>
+        </Box>
+      </Box>
+    </Box>
+  );
+}
